Fix Conv3d dtype mismatch: cast DiT to bf16 before moving to Neuron

+++ inference_neuron_ti2v.py
@@ -117,7 +117,8 @@ def main():
     tp_rank = get_tp_rank()
     shard_model_tp(model, tp_rank, TP_DEGREE)
 
-    # Move sharded model to Neuron
+    # Cast to bf16 then move to Neuron (Conv3d requires matching weight/bias/input dtypes)
+    model = model.to(torch.bfloat16)
     model = model.to(NEURON_DEVICE)
     if rank == 0:
         logger.info("DiT moved to Neuron, compiling sub-modules...")
